Add UM v1.2 codes
Some codes were the same as their v1.1 counterparts, so the existing ones' titles were edited to reflect this, and the codes themselves were moved to the bottom of the list. I consolidated the name of a code as well so that I could differentiate it from another v1.2 code (Fort42 might have cut off the code name because it was too long).